Settings update
se actualizan  los Settings
